% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casabourse.R
\name{construct_download_url}
\alias{construct_download_url}
\title{construct_download_url}
\usage{
construct_download_url(url, format = "csv", sheetid = NULL)
}
\arguments{
\item{url}{link}

\item{format}{a format like csv...}

\item{sheetid}{bool default TRUE}
}
\value{
text
}
\description{
construct_download_url
}
