% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily.data.R
\name{daily.data}
\alias{daily.data}
\title{Daily data}
\usage{
daily.data(ticker = "ATW", from = "28-12-2020", to = "11-08-2022")
}
\arguments{
\item{ticker}{ticker that matches the title that interests us (to remind you know all the tickers of the securities through the tickers() function)}

\item{from}{It represents the start date of data collection. This argument is of the type '' day-month-year ''.}

\item{to}{It represents the date of stopping data collection. This argument is of the type '' day-month-year ''.}
}
\value{
data frame
}
\description{
As its name suggests, this function returns the daily prices of a security between two dates given by the user. It receives three arguments.
}
\examples{
daily.data(ticker = "ATW", from = "01-01-2016", to = "01-01-2021")
}
