% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info.R
\name{get_info}
\alias{get_info}
\title{Get info}
\usage{
get_info(ticker)
}
\arguments{
\item{ticker}{ticker that matches the title that interests us (to remind you know all the tickers of the securities through the tickers() function)}
}
\value{
data frame
}
\description{
This function allows you to get quick information about a company
}
