% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casabourse.R
\name{gsheet2tbl}
\alias{gsheet2tbl}
\title{gsheet2tbl}
\usage{
gsheet2tbl(url, sheetid = NULL)
}
\arguments{
\item{url}{link}

\item{sheetid}{bool default TRUE}
}
\value{
data.frame
}
\description{
It receives a link from a Google sheet spreadsheet as an argument and automatically reads the data that is there in the form of a table without having to download the sheet to your computer.
}
