% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casabourse.R
\name{gsheet2text}
\alias{gsheet2text}
\title{gsheet2txt}
\usage{
gsheet2text(url, format = "csv", sheetid = NULL)
}
\arguments{
\item{url}{link}

\item{format}{a format like csv...}

\item{sheetid}{bool default TRUE}
}
\value{
text
}
\description{
gsheet2txt
}
