% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_pymodules.R
\name{install.tools}
\alias{install.tools}
\title{install.tools}
\usage{
install.tools()
}
\description{
For a first use of the package, you must perform this function that will automatically download all the dependencies you need
}
