% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instruments.R
\name{instruments}
\alias{instruments}
\title{instruments}
\usage{
instruments()
}
\value{
data.frame
}
\description{
It is without argument and gives us information on the financial instruments of the market, such as the ISIN code, the sub-fund, the number of securities etc.
}
\examples{
instruments()
}
