% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pythonFile.R
\name{madex.data}
\alias{madex.data}
\title{Mdex Data}
\usage{
madex.data()
}
\value{
a data.frame
}
\description{
By equivalence to the function masi.data, madex.data returns a data table of the MADEX index.
}
