% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tickers.R
\name{tickers}
\alias{tickers}
\title{tickers}
\usage{
tickers()
}
\value{
data.frame
}
\description{
This function takes no arguments and returns a data table associating each company with a ticker. It is important to visualize the tickers. Indeed, they are used by the other functions of the package to represent the companies to which they correspond.
}
\examples{
tickers()
}
