% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/today.market.R
\name{today.market}
\alias{today.market}
\title{today.market}
\usage{
today.market()
}
\value{
data.frame
}
\description{
is without argument and returns the current price of financial instruments as well as their variations, their opening prices, their max etc.
}
\examples{
today.market()
}
