% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/today.prizelist.R
\name{today.prizelist}
\alias{today.prizelist}
\title{today.prizelist}
\usage{
today.prizelist(up_or_down)
}
\arguments{
\item{up_or_down}{rise or fall prize list}
}
\value{
data.frame
}
\description{
It receives '' up '' or '' down '' respectively and returns a table prizelist of rising or falling prices of market instruments
}
\examples{
\dontrun{today.prizelist('up')}
}
