% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casabourse.R
\name{today.transactions}
\alias{today.transactions}
\title{today.transactions}
\usage{
today.transactions()
}
\value{
data.frame
}
\description{
It allows you to obtain a table of transactions for the day. It is a function without arguments.
}
