% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pythonFile.R
\name{daily.data}
\alias{daily.data}
\title{Title}
\usage{
daily.data(ticker, from, to)
}
\arguments{
\item{ticker}{put ticker}

\item{from}{from date}

\item{to}{to date}
}
\value{
data frame
}
\description{
Title
}
